# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Система бронирования мест в кафе",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1/",
      "description": "Локальный сервер"
    }
  ],
...
  "paths": {
...
    "/dishes": {
      "get": {
        "tags": [
          "Блюда"
        ],
        "summary": "Получение списка блюд",
        "description": "Получение списка блюд. Для администраторов и менеджеров - все блюда (с возможностью выбора), для пользователей - только активные.",
        "operationId": "get_all_dishes_dishes_get",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "show_active",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "title": "Показывать активные блюда?",
              "default": true
            },
            "description": "Показывать все блюда, только активные или только не активные. <br>По умолчанию показывает: <br><ul><li>для пользователя - только активные блюда (всегда, в не зависимости от значения параметра).</li><li>для администратора - все блюда (и активные и не активные)</li><li>для менеджера - активные блюда </li></ul>"
          },
          {
            "name": "cafe_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Cafe Id"
            },
            "description": "ID кафе, в котором показывать блюда. Если не задано - показывает все блюда во всех кафе"
          }
        ],
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DishInfo"
                  },
                  "title": "Возвращает все блюда"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизированный пользователь",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Блюда"
        ],
        "summary": "Создание нового блюда",
        "description": "Создает новое блюда. Только для администраторов и менеджеров.",
        "operationId": "create_dish_dishes_post",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DishCreate",
                "description": "Данные нового блюда"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishInfo"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка в параметрах запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизированный пользователь",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          }
        }
      }
    },
    "/dishes/{dish_id}": {
      "get": {
        "tags": [
          "Блюда"
        ],
        "summary": "Получение информации о блюде по его ID",
        "description": "Получение информации о блюде по его ID. Для администраторов и менеджеров - все блюда, для пользователей - только активные.",
        "operationId": "get_dish_by_id_dishes__dish_id__get",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "dish_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "ID блюда",
              "title": "Dish Id"
            },
            "description": "ID блюда"
          }
        ],
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishInfo"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка в параметрах запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизированный пользователь",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "404": {
            "description": "Данные не найдены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Блюда"
        ],
        "summary": "Обновление информации о блюде по его ID",
        "description": "Обновление информации о блюде по его ID. Только для администраторов и менеджеров.",
        "operationId": "path_dish_by_id_dishes__dish_id__patch",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "dish_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "ID блюда",
              "title": "Dish Id"
            },
            "description": "ID блюда"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DishUpdate",
                "description": "Обновленные данные блюда"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishInfo"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка в параметрах запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизированный пользователь",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "404": {
            "description": "Данные не найдены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          }
        }
      }
    },
...
    "/media/{media_id}": {
      "get": {
        "tags": [
          "Медиа"
        ],
        "summary": "Возвращает изображение в бинарном формате",
        "operationId": "get_media_media__media_id__get",
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "ID изображения",
              "title": "Media Id"
            },
            "description": "ID изображения"
          }
        ],
        "responses": {
          "200": {
            "description": "Успешно. Возвращает изображение в бинарном формате"
          },
          "404": {
            "description": "Данные не найдены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          }
        }
      }
    },
    "/media": {
      "post": {
        "tags": [
          "Медиа"
        ],
        "summary": "Загрузка изображения",
        "description": "Загрузка изображения на сервер. Поддерживаются форматы jpg, png. Размер файла не более 5Мб. Только для администраторов и менеджеров",
        "operationId": "upload_media_media_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/MediaData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaInfo"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка в параметрах запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизированный пользователь",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка сохранения файла",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "MediaData": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File",
            "description": "Загружаемый файл"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "MediaData"
      },
...
      "CafeShortInfo": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "photo_id": {
            "type": "string",
            "format": "uuid",
            "title": "Photo Id"
          }
        },
        "type": "object",
        "title": "CafeShortInfo"
      },
...
      "CustomError": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "CustomError"
      },
      "DishCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "photo_id": {
            "type": "string",
            "format": "uuid",
            "title": "Photo Id"
          },
          "price": {
            "type": "number",
            "title": "Price"
          },
          "cafes_id": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Cafes Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "price",
          "cafes_id"
        ],
        "title": "DishCreate"
      },
      "DishInfo": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "photo_id": {
            "type": "string",
            "format": "uuid",
            "title": "Photo Id"
          },
          "price": {
            "type": "number",
            "title": "Price"
          },
          "cafes": {
            "items": {
              "$ref": "#/components/schemas/CafeShortInfo"
            },
            "type": "array",
            "title": "Cafes"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "title": "DishInfo"
      },
      "DishUpdate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "photo_id": {
            "type": "string",
            "format": "uuid",
            "title": "Photo Id"
          },
          "price": {
            "type": "number",
            "title": "Price"
          },
          "cafes_id": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Cafes Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "DishUpdate"
      },
      "MediaInfo": {
        "properties": {
          "media_id": {
            "type": "string",
            "format": "uuid",
            "title": "Media Id"
          }
        },
        "type": "object",
        "title": "MediaInfo"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}